agent-11-trainer/app/agent.py
# Agent 11 Main - see system_prompt for full implementation
import os
import json
import asyncio
from typing import Dict, List, Any, Optional
from datetime import datetime
import vertexai
from vertexai.generative_models import GenerativeModel

# Initialize
vertexai.init(project=os.getenv("GOOGLE_CLOUD_PROJECT", "your-project-id"))

# Main Agent Definition
class MetaOrchestratorAgent:
    def __init__(self):
        self.model = GenerativeModel("gemini-2.5-pro")
        self.agents = {}
        self.status = "initialized"
    
    async def coordinate(self, request: Dict[str, Any]) -> Dict[str, Any]:
        """Orchestrate multi-agent workflow"""
        try:
            result = await self._execute_workflow(request)
            return result
        except Exception as e:
            return {"error": str(e), "status": "failed"}
    
    async def _execute_workflow(self, request: Dict[str, Any]) -> Dict[str, Any]:
        """Execute coordinated agent workflow"""
        workflow_result = {
            "timestamp": datetime.utcnow().isoformat(),
            "status": "completed",
            "agents_executed": [],
            "results": {}
        }
        return workflow_result

# Create instance
agent_11 = MetaOrchestratorAgent()

if __name__ == "__main__":
    print("Agent 11 - Meta-Agent Orchestrator Ready")
