agent-11-trainer/requirements.txt
google-cloud-aiplatform==1.48.0
google-cloud-vertexai==1.0.2
google-cloud-logging==3.8.0
google-cloud-storage==2.14.0
pydantic==2.5.0
pandas==2.1.3
numpy==1.26.3
asyncio==3.4.3
aiohttp==3.9.1
requests==2.31.0
python-dotenv==1.0.0
